fix: clear renderRightContent cache
